\name{load_events_games}
\alias{load_events_games}
\title{Load Event Games}
\usage{
load_events_games()
}
\description{
Enter event_id to return table of games from selected event. Find unique event_id numbers using load_events_men() or load_events_women(). Table of games includes event name, final scores, tournament stage, and other basic info.
}
\examples{
load_events_games(87)
}
