\name{load_winprob_women}
\alias{load_winprob_women}
\title{Load Women's Win Probability Data}
\usage{
load_winprob_women()
}
\description{
Returns table of DoubleTakeout model's win probability data based on team ranking, score differential, hammer, and ends remaining.

}
\examples{
load_winprob_women()
}
