\name{load_events_stats}
\alias{load_events_stats}
\title{Load Event Stats}
\usage{
load_events_stats()
}
\description{
Enter event_id to return table of overall player stats. Find unique event_id numbers using load_events_men() or load_events_women(). Contains shot quality score, expected quality score, and total throw count by each player for the selected event.

}
\examples{
load_events_stats(251)
}
